pyproject.toml
[project]
name = "b2broker-testapp"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Stanislav Shabalin", email = "services@voxpuibr.com"},
]
dependencies = [
    "djangorestframework>=3.15.2",
    "djangorestframework-jsonapi>=7.0.2",
    "mysqlclient>=2.2.4",
    "django-filter>=24.3",
    "drf-nested-routers>=0.94.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
test = [
    "pytest>=8.3.2",
    "pytest-django>=4.8.0",
    "pytest-cov>=5.0.0",
]
